test: add snmp api test
