+++ typex/version.go
@@ -1,8 +1,8 @@
 //
 // Warning:
 //   This file is generated by go compiler, don't change it!!!
-//   Generated Time: "2022-11-17 14:13:01"
-//   Build on: Linux u22 5.15.0-48-generic #54-Ubuntu SMP Fri Aug 26 13:26:29 UTC 2022 x86_64 x86_64 x86_64 GNU/Linux
+//   Generated Time: "2022-12-08 10:07:34"
+//   Build on: Linux wangwh1-868 4.4.0-17763-Microsoft #2268-Microsoft Thu Oct 07 16:36:00 PST 2021 x86_64 x86_64 x86_64 GNU/Linux
 //
 package typex
 
@@ -12,7 +12,7 @@ type Version struct {
 }
 
 var DefaultVersion = Version{
-	Version:   `v0.4.0`,
-	ReleaseTime: "2022-11-17 14:13:01",
+	Version:   `V0.4.1`,
+	ReleaseTime: "2022-12-08 10:07:34",
 }
